implemented new algorithm
Timestamps are now normalized to session start time 0 and end time 1.
The square of the timestamp is no longer used for the weight of the
first encounter of an edge.  Instead, weights for each edge encounter
are calculated using the exponential function (100^timetamp).  The three
highpass filters used by the algorithm are passed as arguments to
SimpleGraph.exe instead of being hard coded.  SimpleGraph results for
importance are now sorted from highest to lowest score.